fix remarks
