+++ app/components/mobile/object-list-view.js
@@ -1,0 +1,2 @@
+import ObjectListView from 'ember-flexberry/components/mobile/object-list-view';
+export default ObjectListView;
