+++ app/templates/mobile/components/object-list-view.hbs
@@ -113,12 +113,10 @@
   {{#if (not-eq class "groupedit-container")}}
     <div class="ui check-rows-button {{buttonClass}} button {{if readonly 'disabled'}}">
       <i class="check-square-o icon"></i>
-      {{#if checkRowsSettingsItems}}
-        {{flexberry-menu
-          items=checkRowsSettingsItems
-          onItemClick=(action "onCheckRowMenuItemClick")
-        }}
-      {{/if}}
+      {{flexberry-menu
+        items=checkRowsSettingsItems
+        onItemClick=(action "onCheckRowMenuItemClick")
+      }}
     </div>
   {{else}}
     <button
@@ -136,7 +134,7 @@
       class="ui delete-button {{buttonClass}} disabled button"
       title={{t "components.olv-toolbar.delete-button-text"}}
       disabled="disabled"
-      {{action "deleteSelectedRowMobile"}}>
+      {{action "deleteSelectedRow"}}>
         {{t "components.olv-toolbar.delete-button-text"}}
         <i class="delete icon"></i>
     </button>
@@ -144,7 +142,7 @@
     <button
       class="ui delete-button {{buttonClass}} button"
       title={{t "components.olv-toolbar.delete-button-text"}}
-      {{action "deleteSelectedRowMobile"}}>
+      {{action "deleteSelectedRow"}}>
         {{t "components.olv-toolbar.delete-button-text"}}
         <i class="delete icon"></i>
     </button>
@@ -152,7 +150,7 @@
 
   <button class="ui clear-select-button {{buttonClass}} circular icon button"
     title={{t "components.olv-toolbar.clear-select-button-text"}}
-    {{action "clearSelectedRecordsMobile"}}>
+    {{action "clearSelectedRecords"}}>
     <i class="x icon"></i>
   </button>
